Got remote view for workbench working

+++ ION/ION.Droid/src/Fragments/_Workbench/WorkbenchAdapter.cs
@@ -34,30 +34,36 @@ namespace ION.Droid.Fragments._Workbench {
 
 				records.Clear();
 				__workbench = value;
+				__workbench.onWorkbenchEvent += OnWorkbenchEvent;
 
-				foreach (var manifold in workbench.manifolds) {
+				foreach (var manifold in __workbench.manifolds) {
 					var mr = new ManifoldRecord(manifold);
 					records.Add(mr);
 					ExpandManifold(mr);
+					records.Add(new SpaceRecord());
 				}
 
 				records.Add(new AddViewerRecord(onAddViewer));
 
-				if (__workbench != null) {
-					__workbench.onWorkbenchEvent += OnWorkbenchEvent;
-				}
-
 				NotifyDataSetChanged();
+
 			}
 		} Workbench __workbench;
+
+		/// <summary>
+		/// Whether or not the adapter is user editable.
+		/// </summary>
+		/// <value><c>true</c> if is editable; otherwise, <c>false</c>.</value>
+		public bool isEditable { get; private set; } bool __isEditable;
 
 		private BitmapCache cache;
 		private Action onAddViewer;
 		private ItemTouchHelper dragger;
 		private HashSet<ManifoldRecord> dragCollapsedManifoldRecords = new HashSet<ManifoldRecord>();
 
-		public WorkbenchAdapter(Action onAddViewer) {
+		public WorkbenchAdapter(Action onAddViewer, bool isEditable) {
 			this.onAddViewer = onAddViewer;
+			this.isEditable = isEditable;
 			dragger = new ItemTouchHelper(new Dragger(this));
 		}
 
@@ -66,12 +72,22 @@ namespace ION.Droid.Fragments._Workbench {
 			if (cache == null) {
 				cache = new BitmapCache(recyclerView.Context.Resources);
 			}
-			dragger.AttachToRecyclerView(recyclerView);
+			if (isEditable) {
+				dragger.AttachToRecyclerView(recyclerView);
+			}
+/*
+			var anim = recyclerView.GetItemAnimator() as SimpleItemAnimator;
+			if (anim != null) {
+				anim.SupportsChangeAnimations = false;
+			}
+*/
 		}
 
 		public override void OnDetachedFromRecyclerView(RecyclerView recyclerView) {
 			base.OnDetachedFromRecyclerView(recyclerView);
-			dragger.AttachToRecyclerView(null);
+			if (isEditable) {
+				dragger.AttachToRecyclerView(null);
+			}
 		}
 
 		// Overridden from RecordAdapter
@@ -85,6 +101,9 @@ namespace ION.Droid.Fragments._Workbench {
 			var rv = recyclerView as SwipeRecyclerView;
 
 			switch ((EViewType)viewType) {
+				case EViewType.Space: {
+					return new SpaceViewHolder(parent);
+				} // EViewType.Space
 				case EViewType.Add: {
 					return new AddViewerViewHolder(parent);
 				} // EViewType.Add
@@ -233,7 +252,7 @@ namespace ION.Droid.Fragments._Workbench {
 		/// <returns>The size for manifold.</returns>
 		/// <param name="manifold">Manifold.</param>
 		private int AdapterSizeForManifold(ManifoldRecord mr) {
-			return 1 + (mr.isExpanded ? mr.manifold.sensorPropertyCount : 0);
+			return 1 + (mr.isExpanded ? mr.manifold.sensorPropertyCount : 0) + 1;
 		}
 
 		private IRecord CreateSensorPropertyRecord(Manifold manifold, ISensorProperty sp) {
@@ -293,7 +312,9 @@ namespace ION.Droid.Fragments._Workbench {
 
 		private void PerformSwap(RecyclerView.ViewHolder vh1, RecyclerView.ViewHolder vh2) {
 			if (vh1 is ManifoldViewHolder && vh2 is ManifoldViewHolder) {
-				workbench.Swap(vh1.AdapterPosition, vh2.AdapterPosition);
+				var m1 = ((ManifoldViewHolder)vh1).record.manifold;
+				var m2 = ((ManifoldViewHolder)vh2).record.manifold;
+				workbench.Swap(workbench.IndexOf(m1), workbench.IndexOf(m2));
 			} else if (vh1 is SensorPropertyViewHolder && vh2 is SensorPropertyViewHolder) {
 				var spr1 = records[vh1.AdapterPosition] as SensorPropertyRecord;
 				var spr2 = records[vh2.AdapterPosition] as SensorPropertyRecord;
@@ -310,12 +331,20 @@ namespace ION.Droid.Fragments._Workbench {
 						var i = 0;
 						var mr = new ManifoldRecord(e.manifold);
 						if (e.index == 0) {
+							records.Insert(i, new SpaceRecord());
 							records.Insert(i, mr);
+							NotifyItemRangeInserted(i, 2);
 						} else {
 							i = AdapterIndexForManifold(workbench[e.index - 1]);
+							i += AdapterSizeForManifold(records[i] as ManifoldRecord);
+							records.Insert(i, new SpaceRecord());
 							records.Insert(i, mr);
+							NotifyItemRangeInserted(i, 2);
 						}
-						NotifyItemInserted(i);
+
+						if (mr.manifold.sensorPropertyCount > 0) {
+							ExpandManifold(mr);
+						}
 					}
 				break; // WorkbenchEvent.EType.Added
 				case WorkbenchEvent.EType.ManifoldEvent: {
@@ -323,9 +352,11 @@ namespace ION.Droid.Fragments._Workbench {
 				} break; // WorkbenchEvent.EType.ManifoldEvent
 				case WorkbenchEvent.EType.Removed: {
 					var i = AdapterIndexForManifold(e.manifold);
-					var cnt = AdapterSizeForManifold(records[i] as ManifoldRecord);
-					records.RemoveRange(i, cnt);
-					NotifyItemRangeRemoved(i, cnt);
+					if (i != -1) {
+						var cnt = AdapterSizeForManifold(records[i] as ManifoldRecord);
+						records.RemoveRange(i, cnt);
+						NotifyItemRangeRemoved(i, cnt);
+					}
 				} break; // WorkbenchEvent.EType.Removed
 				case WorkbenchEvent.EType.Swapped: {
 					SwapRecords(AdapterIndexForManifold(workbench[e.index]), AdapterIndexForManifold(workbench[e.otherIndex]));
@@ -338,6 +369,9 @@ namespace ION.Droid.Fragments._Workbench {
 				case ManifoldEvent.EType.SensorPropertyAdded: {
 					var aifm = AdapterIndexForManifold(e.manifold);
 					var mr = records[aifm] as ManifoldRecord;
+					if (mr.manifold.sensorPropertyCount == 1) {
+						mr.isExpanded = true;
+					}
 					if (mr.isExpanded) {
 						var i = aifm + e.index + 1;
 						var record = CreateSensorPropertyRecord(e.manifold, e.manifold[e.index]);
@@ -367,11 +401,12 @@ namespace ION.Droid.Fragments._Workbench {
 					var i1 = mi + e.index;
 					var i2 = mi + e.otherIndex;
 					SwapRecords(i1, i2);
-
+/*
 					var vh1 = recyclerView.FindViewHolderForAdapterPosition(i1) as SensorPropertyViewHolder;
 					var vh2 = recyclerView.FindViewHolderForAdapterPosition(i2) as SensorPropertyViewHolder;
 					vh1.Invalidate();
 					vh2.Invalidate();
+*/
 				} break; // ManifoldEvent.EType.SensorPropertySwapped
 			}
 		}
@@ -389,6 +424,8 @@ namespace ION.Droid.Fragments._Workbench {
 		}
 
 		public enum EViewType {
+			Space,
+
 			Add,
 			Manifold,
